Removed BigRational dependency from VefParser
I moved the number parsing behind the AlgebraicField interface.

All unit tests are passing, including the big number VEF.

+++ core/src/main/java/com/vzome/core/viewing/Camera.java
@@ -142,7 +142,7 @@ public class Camera
 
     public float getViewDistance()
     {
-        return (float) mDistance;
+        return mDistance;
     }
 
     public void getViewOrientation( Vector3f lookDir, Vector3f upDir )
